fix: data types matter folks...
This fixes an issue where we compare floats to 0 (int) but technically floats need to be compared to 0.0. In this case we don't particularly care the about the float itself we just need to figure out if there is an offset worth dealing with. casting to int is just fine.

+++ app/Domain/Timesheets/Services/Timesheets.php
@@ -305,7 +305,7 @@ class Timesheets
             }
             // Detect timezone offset
 
-            $workdateOffsetStart = ($currentWorkDate->setToUserTimezone()->secondsSinceMidnight() / 60 / 60);
+            $workdateOffsetStart = (int) ($currentWorkDate->setToUserTimezone()->secondsSinceMidnight() / 60 / 60);
 
             // Various Entries can be in different timezones and thus would not be caught by upsert or grouping by
             // default Creating new rows for each timezone adjustment
